Module tarpc::macros::serde::de

Deserialization re-exports required by macros. Not for general use.

Traits

EnumVisitor
EnumVisitor is a visitor that is created by the Deserialize and passed to the Deserializer in order to deserialize enums.

Error
Error is a trait that allows a Deserialize to generically create a Deserializer error.

VariantVisitor
VariantVisitor is a visitor that is created by the Deserializer and passed to the Deserialize in order to deserialize a specific enum variant.

Visitor
This trait represents a visitor that walks through a deserializer.