Module tarpc::protocol

Provides the tarpc client and server, which implements the tarpc protocol. The protocol is defined by the implementation.

Structs

Client
A client stub that connects to a server to run rpcs.

Future
An asynchronous RPC call

ServeHandle
Provides methods for blocking until the server completes,

Enums

Error
Client errors that can occur during rpc calls

Traits

Serve
A service provided by a server

Functions

serve_async
Start

Type Definitions

Result
Return type of rpc calls: either the successful return value, or a client error.